Changed latex command to a cog
