+++ shard/res/generate/template.tex
